Fix 64-bit width bugs
High-priority fixes:
- #1: megamorphic lookup-cache probe regenerated with real cacheElement stride;
  cache addresses widened to intptr_t (AArch64 materializes base in register,
  probes absolute; x86-64 same via movabs into edi).
- #2: nlr_home/nlr_home_id widened to intptr_t across abort.cpp, process.cpp/hpp,
  stubRoutines.cpp; x86-64 NLR stub uses movq/notq to keep 64-bit values.
- #3: scopeDescRecorder oops array now intptr_t-based; getValueIndex(intptr_t);
  added limits.h for min_jint on LP64.
- #4: interpreter dispatch stride now times_8 on x86-64; call_C/PRIM_ARG_DECL
  stack offsets corrected; rest tracked under item 5 (PICs/stubs/asm).

Medium batch (width/formatting/one-liners):
- #5: unsigned_mul already fixed (confirmed); isExternalWord corrected.
- #6: nmethodScopes::getNext boundary now oops_offset() (bytes, not scaled).
- #9: relocInfo backend-selected: quad_addr() for AArch64 8-byte literals,
  word_addr() for x86 disp32; fix_relocation_at_move selects per backend;
  callDestination/call_end dispatch width corrected; disasm print target fixed.
- #10: dispatchTable intercept_for_return now uses dispatch_table[code];
  frame_breakpoint sentinel changed to (void*)-1.
- #11: SMI_RELATIONAL_OP now intptr_t; bitShift rewritten with BitsPerWord=32
  and intptr_t masks (no signed-int UB). smi_min/smi_max NOT widened
  (snapshot/29-bit-smi compat).
- #12: double_prims six as_smiOop truncations fixed (int -> intptr_t).
- #15: disassembler libname buffer enlarged (13 -> 32 bytes).
- #16: floatSize asserts accept oopSize or 2*oopSize on both backends.
- #17: inline.cpp dead (ifdef COMPILER never defined) - no action.
- #18: tag.hpp clearTag() now intptr_t cast.
- #19: nmethod::isFree() reads _instsLen as full pointer word; primitives::lookup
  rewritten as bounds-safe binary search; integerOps print left (debug-only).

+++ vm/code/scopeDescRecorder.hpp
@@ -149,7 +149,7 @@ private:
 
   void updateExtScopeDescHeader(int offset, int next);
 
-  inline int getValueIndex(int v);
+  inline int getValueIndex(intptr_t v);
   inline int getOopIndex(oop o);
 
   inline void genIndex(int index);
